update redux store folder structure and unit tests
