+++ src/store/types/sidebar.js
@@ -1,0 +1,3 @@
+export const OPEN_SIDEBAR = "OPEN_SIDEBAR";
+export const CLOSE_SIDEBAR = "CLOSE_SIDEBAR";
+export const TOGGLE_SIDEBAR = "TOGGLE_SIDEBAR";
